netatalk: tighten single-user mode and the daemonize logging path
daemonize() now owns the logger's descriptors: it closes the syslog
connection and every log file before closeall(), so no stale number is
left for log_reopen() to close after another type has been given it,
and reopens every configured log after the standard descriptors are
/dev/null again. A root controller with the default syslog logging
opened its connection during the config parse, lost it to closeall()
and, on musl, then dropped every later line for the life of the
process; a 'log file' in mkstemp form kept a stale descriptor the same
way. The logger stores the name a file was opened by, the mkstemp name
included, as an allocated absolute path, so a relative 'log file' is
reopened in the launch directory after daemonize() has moved to /; a
file whose name cannot be kept is not set up at all. The controller no
longer reopens on its own.

afpd judged "not logged in" by euid 0, which a server started without
root never has, so a client that opened a DSI session and dropped
before login was parked for 'disconnect time' instead of ending its
child, and afp_dsi_close() tried seteuid(0). login() sets obj->logout,
and that is what afp_alarm(), dsi_disconnect() and afp_dsi_close()
test now.

readvolfile() counts a volume section without 'path', with a log line
naming it, and the two volxlate() failures, so the controller's
"every configured volume must load" clause sees them; a [Homes]
section is refused by its presence, not only when it carries a
'basedir regex', and so is 'vol dbnest', which would make each volume
root its own CNID directory and have the owner-only rule chmod it
0700. The CNID directory clause drops trailing slashes
before it looks, so the directory a [Global] 'vol dbpath' derives,
which ends in one, is refused as a symbolic link like the volume-level
form; a trailing slash makes lstat() follow the link.

cnid_sqlite_open() creates an absent CNID directory shared whoever the
opener is, and keeps state owner-only for the single-user flag or for
a directory the non-root opener already holds as its own 0700
directory: the serving user's nad before the server's first start now
leaves shared state the server tightens, and a non-root nad on a
shared server no longer creates owner-only state other users cannot
open. The message for a directory that cannot be opened says why.

afppasswd's private bootstrap opens the verifier O_EXCL first and only
removes on failure a file this call created, so an -f run that
overlaps another cannot unlink the other run's verifier.

The controller splits 'uam list' the way auth_load() does, tests each
volume's access with one access() call, derives a parent directory
without dirname(3) (illumos declares bgets() in <libgen.h>, which
collides with bstrlib's), and sorts its includes; the afptest client's
includes are sorted, its gcrypt header sits in its own cluster, and its
SRP helpers document every parameter.

netatalk.8 and NEWS say that without --single-user netatalk requires
root and exits 1; Configuration.md says which root-created CNID files
a single-user server cannot tighten. The container entrypoint probes
the configured AFP_HOST, and the single-user CI leg no longer dumps the
server log on success.

singleuser_test.sh gains legs for a section without a path, a [Homes]
section without a basedir regex, 'vol dbnest', a symbolic link under a
[Global] vol dbpath, a connection dropped before login, and shared
state nad creates before the first start, and asserts the CNID
directory stays owner-only once the desktop database exists;
logreopen_test.sh gains the syslog, mkstemp and relative-name cases.

+++ test/afppasswd/test_afppasswd.c
@@ -2,6 +2,7 @@
  * Tests for afppasswd's SRP verifier validation
  *
  * Copyright (c) 2026 Daniel Markstedt <daniel@mindani.net>
+ * Copyright (c) 2026 Andy Lemin (andylemin)
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -673,7 +674,7 @@ static int read_single_record(const char *path, char *line, size_t size)
         return 0;
     }
 
-    single = fgets(line, size, fp) != NULL && fgetc(fp) == EOF;
+    single = fgets(line, (int)size, fp) != NULL && fgetc(fp) == EOF;
     fclose(fp);
     return single;
 }
